requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pytest>=6.2.5
httpx>=0.19.0  # For TestClient in integration tests
pytest-playwright>=0.3.0  # For E2E tests
python-multipart>=0.0.5